fix(build): Moving from black to ruff for test format and import checks
Because of the speed of ruff and its closeness to black style,
we decided to use ruff for autoformatting and linting on this
project. The minor differences between ruff and black make it
impractical to retain black in the pytest checks. So we're
switching the pytest setup to use ruff as well. While we're at
it we're also switching from isort to ruff for import sorting.
This should speed up test runs noticeably.