wip(aggregations): Disabling by default the extraneous aggregations
The resource overhead for performing catchup aggregations on all three
date bases is too much for the present, given that we only plan to
make the "added" basis stats visible at present. So we've commented out
the others in the key config variables. Enabling them again will be
a simple matter of uncommenting those items.